still a little error in the triangulation, but getting closer
added some properties for rendering (CullMode, FillMode => rasterizer states and UseChache flag)

+++ Projects/Transvoxel/Geometry/Mesh.cs
@@ -68,9 +68,9 @@ namespace Transvoxel.Geometry
             return arr;
         }
 
-        internal int VertexCount()
+        internal ushort VertexCount()
         {
-            return vertices.Count;
+            return (ushort)vertices.Count;
         }
     }
 }
